requirements.pip.txt
flask-basicauth
imageio
nose-exclude
solrpy
